+++ app/ante/evm/interfaces.go
@@ -26,6 +26,8 @@ type EVMKeeper interface { //nolint: revive
 	SetupExecutionContext(ctx sdk.Context, txGas uint64, txType uint8) sdk.Context
 	GetTxCountTransient(ctx sdk.Context) uint64
 	GetParams(ctx sdk.Context) evmtypes.Params
+
+	SetFlagSenderNonceIncreasedByAnteHandle(ctx sdk.Context, increased bool)
 }
 
 type FeeMarketKeeper interface {