+++ app/ante/evm/setup_ctx.go
@@ -40,6 +40,9 @@ func (esc EthSetupContextDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simul
 	newCtx = ctx.WithGasMeter(storetypes.NewInfiniteGasMeter())
 	newCtx = utils.UseZeroGasConfig(newCtx)
 
+	// reset previous run
+	esc.evmKeeper.SetFlagSenderNonceIncreasedByAnteHandle(newCtx, false)
+
 	return next(newCtx, tx, simulate)
 }