+++ x/evm/types/key.go
@@ -40,6 +40,7 @@ const (
 	prefixTransientTxGas
 	prefixTransientTxLogCount
 	prefixTransientTxReceipt
+	prefixTransientFlagIncreasedSenderNonce
 )
 
 // KVStore key prefixes
@@ -59,7 +60,8 @@ var (
 
 // Transient Store key
 var (
-	KeyTransientTxCount = []byte{prefixTransientTxCount}
+	KeyTransientTxCount                  = []byte{prefixTransientTxCount}
+	KeyTransientFlagIncreasedSenderNonce = []byte{prefixTransientFlagIncreasedSenderNonce}
 )
 
 // AddressStoragePrefix returns a prefix to iterate over a given account storage.